+++ src/core/interpreter/http/utils.ts
@@ -142,7 +142,5 @@ export async function fetchRequest(
 }
 
 export function hasAcceptHeader(headers: NonPrimitive): boolean {
-  return Object.keys(headers).some(
-    header => header.toLowerCase() === 'accept'
-  );
+  return Object.keys(headers).some(header => header.toLowerCase() === 'accept');
 }